+++ src/callback.rs
@@ -20,6 +20,8 @@ pub struct Callback48<T> {
     [ f32 ]   [ Callback32 ];
 )]
 impl<T> Callback48<T> {
+    /// Create a new callback from a closure. Time between updates is indicated by `update_interval`,
+    /// which is in seconds.
     pub fn new(update_interval: f48, callback: Box<dyn FnMut(f48, f48, &mut T) + Send>) -> Self {
         Self {
             callback,
@@ -28,10 +30,14 @@ impl<T> Callback48<T> {
             delta_time: 0.0,
         }
     }
+
+    /// Reset time to zero.
     pub fn reset(&mut self) {
         self.time = 0.0;
         self.delta_time = 0.0;
     }
+
+    /// Indicate that time is about to elapse.
     pub fn update(&mut self, dt: f48, object: &mut T) {
         // The first update is always done at time zero.
         if self.delta_time >= self.update_interval || (self.time == 0.0 && dt > 0.0) {
@@ -40,5 +46,15 @@ impl<T> Callback48<T> {
         }
         self.delta_time += dt;
         self.time += dt;
+    }
+
+    /// Set current time. The next update will be issued at the new time.
+    pub fn set_time(&mut self, time: f48) {
+        self.time = time;
+        if time > 0.0 {
+            self.delta_time = self.update_interval;
+        } else {
+            self.delta_time = 0.0;
+        }
     }
 }